+++ clouddns/src/test/java/denominator/clouddns/CloudDNSProviderTest.java
@@ -2,8 +2,8 @@ package denominator.clouddns;
 
 import static denominator.CredentialsConfiguration.credentials;
 import static denominator.Denominator.create;
-import static denominator.Denominator.providers;
-import static denominator.Denominator.provider;
+import static denominator.Providers.list;
+import static denominator.Providers.provide;
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertTrue;
 
@@ -34,7 +34,7 @@ public class CloudDNSProviderTest {
 
     @Test
     public void testCloudDNSRegistered() {
-        Set<Provider> allProviders = ImmutableSet.copyOf(providers());
+        Set<Provider> allProviders = ImmutableSet.copyOf(list());
         assertTrue(allProviders.contains(PROVIDER));
     }
 
@@ -58,7 +58,7 @@ public class CloudDNSProviderTest {
     @Test
     public void testViaDagger() {
         DNSApiManager manager = ObjectGraph
-                .create(provider(new CloudDNSProvider()), new CloudDNSProvider.Module(), credentials("username", "apiKey"))
+                .create(provide(new CloudDNSProvider()), new CloudDNSProvider.Module(), credentials("username", "apiKey"))
                 .get(DNSApiManager.class);
         assertEquals(manager.api().zones().getClass(), CloudDNSZoneApi.class);
     }