+++ core/src/main/java/denominator/Denominator.java
@@ -1,18 +1,9 @@
 package denominator;
 
-import static denominator.common.Preconditions.checkArgument;
-import static denominator.common.Preconditions.checkNotNull;
-
-import java.lang.reflect.Constructor;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.ServiceLoader;
 
-import javax.inject.Singleton;
-
-import dagger.Module;
 import dagger.ObjectGraph;
-import dagger.Provides;
 
 /**
  * Entry-point to create instances of {@link DNSApiManager}
@@ -39,7 +30,7 @@ import dagger.Provides;
  * 
  * <pre>
  * import static denominator.CredentialsConfiguration.credentials;
- * import static denominator.Denominator.provider;
+ * import static denominator.Providers.provide;
  * 
  * ...
  * 
@@ -66,20 +57,11 @@ public final class Denominator {
     }
 
     /**
-     * Returns the currently configured {@link Provider providers} from
-     * {@link ServiceLoader#load(Class)}.
-     * 
-     * <br>
-     * <br>
-     * <b>Performance Note</b><br>
-     * 
-     * The implicit call {@link ServiceLoader#load(Class)} can add delays
-     * measurable in 10s to hundreds of milliseconds depending on the number of
-     * jars in your classpath. If you desire speed, it is best to instantiate
-     * Providers directly.
+     * @deprecated use {@link Providers#list()}. to be removed in denominator 4.
      */
+    @Deprecated
     public static Iterable<Provider> providers() {
-        return ServiceLoader.load(Provider.class);
+        return Providers.list();
     }
 
     /**
@@ -115,42 +97,12 @@ public final class Denominator {
 
     private static List<Object> modulesForGraph(Provider in, Object... modules) {
         List<Object> modulesForGraph = new ArrayList<Object>(3);
-        modulesForGraph.add(provider(in));
-        modulesForGraph.add(instantiateModule(in));
+        modulesForGraph.add(Providers.provide(in));
+        modulesForGraph.add(Providers.instantiateModule(in));
         if (modules != null)
             for (Object module : modules)
                 modulesForGraph.add(module);
         return modulesForGraph;
-    }
-
-    private static Object instantiateModule(Provider in) throws IllegalArgumentException {
-        String moduleClassName;
-        if (in.getClass().isAnonymousClass()) {
-            moduleClassName = in.getClass().getSuperclass().getName() + "$Module";
-        } else {
-            moduleClassName = in.getClass().getName() + "$Module";
-        }
-        Class<?> moduleClass;
-        try {
-            moduleClass = Class.forName(moduleClassName);
-        } catch (ClassNotFoundException e) {
-            throw new IllegalArgumentException(in.getClass().getName()
-                    + " should have a static inner class named Module", e);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("exception attempting to instantiate " + moduleClassName
-                    + " for provider " + in.name(), e);
-        }
-        try {
-            Constructor<?> ctor = moduleClass.getDeclaredConstructor();
-            // allow private or package protected ctors
-            ctor.setAccessible(true);
-            return ctor.newInstance();
-        } catch (NoSuchMethodException e) {
-            throw new IllegalArgumentException("ensure " + moduleClassName + " has a no-args constructor", e);
-        } catch (Exception e) {
-            throw new IllegalArgumentException("exception attempting to instantiate " + moduleClassName
-                    + " for provider " + in.name(), e);
-        }
     }
 
     /**
@@ -170,67 +122,15 @@ public final class Denominator {
      *             {@link Provider} is misconfigured or expects credentials.
      */
     public static DNSApiManager create(String providerName, Object... modules) throws IllegalArgumentException {
-        checkNotNull(providerName, "providerName");
-        Provider matchedProvider = null;
-        List<String> providerNames = new ArrayList<String>();
-        for (Provider provider : providers()) {
-            if (provider.name().equals(providerName)) {
-                matchedProvider = provider;
-                break;
-            }
-            providerNames.add(provider.name());
-        }
-        checkArgument(matchedProvider != null, "provider %s not in set of configured providers: %s", providerName,
-                providerNames);
+        Provider matchedProvider = Providers.getByName(providerName);
         return create(matchedProvider, modules);
     }
 
     /**
-     * Use this when building {@link DNSApiManager} via Dagger.
-     * 
-     * ex. when no runtime changes to the provider are necessary:
-     * 
-     * <pre>
-     * ultraDns = ObjectGraph.create(provider(new UltraDNSProvider()),
-     *                               new UltraDNSProvider.Module(),
-     *                               credentials(username, password)).get(DNSApiManager.class);
-     * </pre>
-     * 
-     * ex. for dynamic provider
-     * 
-     * <pre>
-     * Provider fromDiscovery = new UltraDNSProvider() {
-     *     public String getUrl() {
-     *         return discovery.getUrlFor(&quot;ultradns&quot;);
-     *     }
-     * };
-     * 
-     * ultraDns = ObjectGraph.create(provider(fromDiscovery),
-     *                               new UltraDNSProvider.Module(),
-     *                               credentials(username, password)).get(DNSApiManager.class);
-     * </pre>
+     * @deprecated use {@link Providers#provide}. to be removed in denominator
+     *             4.
      */
     public static Object provider(denominator.Provider provider) {
-        return new ProvideProvider(provider);
-    }
-
-    @Module(injects = DNSApiManager.class, complete = false)
-    static final class ProvideProvider implements javax.inject.Provider<Provider> {
-        private final denominator.Provider provider;
-
-        private ProvideProvider(denominator.Provider provider) {
-            this.provider = checkNotNull(provider, "provider");
-        }
-
-        @Provides
-        @Singleton
-        public Provider get() {
-            return provider;
-        }
-
-        @Override
-        public String toString() {
-            return "Provides(" + provider + ")";
-        }
+        return Providers.provide(provider);
     }
 }